+++ liga/src/de/liga/dart/fileimport/vfs/LITSAD.java
@@ -9,8 +9,8 @@ import java.sql.Date;
  */
 class LITSAD {
     Date SAI_NR;  // Saison Nr (Date), z.b. 01.01.2009
-    long LIG_NR; // Ligaklase.PK
+    int LIG_NR; // Ligaklase.PK
     int SAI_POSNR; // 1-8
-    long TEA_NR;   // TEA_NR,N,5,0 Team.PK
+    int TEA_NR;   // TEA_NR,N,5,0 Team.PK
     String SPI_NR; // P0 fix
 }
