+++ liga/src/de/liga/dart/fileimport/vfs/LITTEA.java
@@ -9,12 +9,12 @@ import java.sql.Date;
  */
 class LITTEA {
     Date SAI_NR; 
-    long LIG_NR;   // Ligaklasse
-    long TEA_NR;   // TEA_NR,N,5,0 PK
+    int LIG_NR;   // Ligaklasse
+    int TEA_NR;   // TEA_NR,N,5,0 PK
     String TEA_NAME;
-    long LOK_NR;   // Spielort.externe ID
+    int LOK_NR;   // Spielort.externe ID
     //String TEA_KAPIT;
     String TEA_SPIELT; // Freitag ...
-    long TEA_UHRZEI; // N(5) 72000 (20:00) 68400 (:60:60 = 19:00)
+    int TEA_UHRZEI; // N(5) 72000 (20:00) 68400 (:60:60 = 19:00)
     String TEA_STATUS; // D = Spielfrei
 }
